requirements.txt
camelot-py[cv]
pydantic
PyMuPDF
streamlit
pandas
fuzzywuzzy
python-levenshtein